+++ BlogposterCMS/mother/modules/databaseManager/meltdownBridging/createDatabaseEvent.js
@@ -4,11 +4,7 @@
  * meltdown => "createDatabase"
  */
 const { getEngine } = require('../engines/engineFactory');
-const { moduleHasOwnDb } = require('../helpers/dbTypeHelpers');
-const { createOrFixSchemaInMainDb } = require('../engines/postgresEngine');
-const { createMongoDatabase } = require('../engines/mongoEngine');
-const { createOrFixSqliteDatabaseForModule } = require('../engines/sqliteEngine');
-const { getDbType } = require('../helpers/dbTypeHelpers');
+const { moduleHasOwnDb, getDbType } = require('../helpers/dbTypeHelpers');
 const { onceCallback } = require('../../../emitters/motherEmitter');
 
 // NEW: Notification Emitter for typed notifications
@@ -50,7 +46,7 @@ function registerCreateDatabaseEvent(motherEmitter) {
           callback(null, { success: true, type: 'ownDb' });
         } else {
           // create or fix a shared schema in main DB
-          await createOrFixSchemaInMainDb(moduleName);
+          await engine.createOrFixSchemaInMainDb(moduleName);
           notificationEmitter.notify({
             moduleName: 'databaseManager',
             notificationType: 'info',
@@ -60,7 +56,7 @@ function registerCreateDatabaseEvent(motherEmitter) {
           callback(null, { success: true, type: 'sharedSchema' });
         }
       } else if (dbType === 'mongodb') {
-        await createMongoDatabase(moduleName);
+        await engine.createMongoDatabase(moduleName);
         notificationEmitter.notify({
           moduleName: 'databaseManager',
           notificationType: 'info',
@@ -69,7 +65,7 @@ function registerCreateDatabaseEvent(motherEmitter) {
         });
         callback(null, { success: true, type: 'mongodb' });
       } else if (dbType === 'sqlite') {
-        await createOrFixSqliteDatabaseForModule(moduleName, isOwnDb);
+        await engine.createOrFixSqliteDatabaseForModule(moduleName, isOwnDb);
         notificationEmitter.notify({
           moduleName: 'databaseManager',
           notificationType: 'info',